Move vtk graph specific tests to VTK plugin
