+++ romanesco/plugins/vtk/plugin.cmake
@@ -1,0 +1,1 @@
+add_python_test(graph PLUGIN vtk PLUGINS_ENABLED vtk)
